feat: implement patient dashboard and role-based root redirection (fix #104)
